Added a proposal of the new Tile object.
This is just a proposal of the new Tile object, an may be changed at any time.
Allowing it to return an integer on onWalk lets us easier do tasks like looking for wild pokemon, from anywhere in the program.
It cannot be used as it is now, but it can be used once we are finished with the other necessary classes.